pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video2prompt"
version = "0.1.0"
description = "本地批量抖音视频解析与 Gemini 解读工具"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "streamlit>=1.39.0",
  "httpx>=0.27.0",
  "aiosqlite>=0.20.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.1",
  "openpyxl>=3.1.5"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "respx>=0.21.1"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
